+++ pyrax/identity/keystone_identity.py
@@ -18,7 +18,7 @@ class KeystoneIdentity(BaseIdentity):
     _default_region = "RegionOne"
 
     def _get_auth_endpoint(self):
-        ep = pyrax.get_setting("auth_endpoint")
+        ep = self._auth_endpoint or pyrax.get_setting("auth_endpoint")
         if ep is None:
             raise exc.EndpointNotDefined("No auth endpoint has been specified.")
         return ep